Alternate Triangle Gen Logic
This triangle generation logic flips the orientation of the hull triangles, to remove concavity, and changes the stern and bow generation to be more low-poly friendly. This is not an all-good change, and can lead to a different kind of artifacts. Concavity is still sometimes present on the back, but this version works much better at generating low-poly models, and in certain low settings, has NO visible artifacts. This is arguably better in general than the previous version, but perhaps a more perfect solution will be found.